Prevent integer overflow in kernel and userland when checking mbuf limits.  Convert kernel variables and calculations for mbuf memory into long to allow larger values on 64 bit machines.  Put a range check into the kernel sysctl.  For the interface itself int is still sufficient.  In netstat -m cast all multiplications to unsigned long to hold the product of two unsigned int. input and OK visa@

+++ sys/conf/param.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: param.c,v 1.40 2019/04/28 14:51:16 deraadt Exp $	*/
+/*	$OpenBSD: param.c,v 1.41 2019/07/08 23:59:32 mlarkin Exp $	*/
 /*	$NetBSD: param.c,v 1.16 1996/03/12 03:08:40 mrg Exp $	*/
 
 /*
@@ -89,7 +89,7 @@ int	initialvnodes = NVNODE;
 int	maxprocess = NPROCESS;
 int	maxthread = NPROCESS + 8 * MAXUSERS;
 int	maxfiles = 5 * (NPROCESS + MAXUSERS) + 80;
-int	nmbclust = NMBCLUSTERS;
+long	nmbclust = NMBCLUSTERS;
 
 #ifndef MBLOWAT
 #define MBLOWAT		16
